refactor: introduce in-memory vector store
