+++ compact_memory/__init__.py
@@ -10,8 +10,8 @@ __all__ = [
     "BeliefPrototype",
     "RawMemory",
     "ConversationalTurn",
-    "JsonNpyVectorStore",
-    "VectorStore",
+    "InMemoryVectorStore",
+    "BaseVectorStore",
     "Agent",
     "QueryResult",
     "PrototypeHit",
@@ -53,8 +53,8 @@ _lazy_map = {
     "BeliefPrototype": "compact_memory.models",
     "RawMemory": "compact_memory.models",
     "ConversationalTurn": "compact_memory.models",
-    "JsonNpyVectorStore": "compact_memory.json_npy_store",
-    "VectorStore": "compact_memory.json_npy_store",
+    "InMemoryVectorStore": "compact_memory.vector_store",
+    "BaseVectorStore": "compact_memory.vector_store",
     "Agent": "compact_memory.agent",
     "QueryResult": "compact_memory.agent",
     "PrototypeHit": "compact_memory.agent",
