--- compact_memory/json_npy_store.py
@@ -1,252 +1,0 @@
-from __future__ import annotations
-
-import json
-import os
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import Dict, List, Tuple
-
-import faiss
-
-import numpy as np
-import yaml
-
-from .models import BeliefPrototype, RawMemory
-from .embedding_pipeline import EmbeddingDimensionMismatchError
-
-
-class VectorStore:
-    """Abstract storage interface."""
-
-    def add_prototype(self, proto: BeliefPrototype, vec: np.ndarray) -> None:
-        raise NotImplementedError
-
-    def update_prototype(
-        self, proto_id: str, new_vec: np.ndarray, memory_id: str
-    ) -> None:
-        raise NotImplementedError
-
-    def find_nearest(self, vec: np.ndarray, k: int) -> List[Tuple[str, float]]:
-        raise NotImplementedError
-
-    def add_memory(self, memory: RawMemory) -> None:
-        raise NotImplementedError
-
-    def save(self) -> None:
-        raise NotImplementedError
-
-    def load(self) -> None:
-        raise NotImplementedError
-
-
-class JsonNpyVectorStore(VectorStore):
-    """Filesystem-backed store using JSON + NPY files."""
-
-    def __init__(
-        self,
-        path: str,
-        *,
-        embedding_model: str = "unknown",
-        embedding_dim: int = 0,
-        normalized: bool = True,
-    ) -> None:
-        self.path = Path(path)
-        self.embedding_model = embedding_model
-        self.embedding_dim = embedding_dim
-        self.normalized = normalized
-        self.meta: Dict[str, object] = {}
-        self.prototypes: List[BeliefPrototype] = []
-        self.proto_vectors: np.ndarray | None = None
-        self.memories: List[RawMemory] = []
-        self.index: Dict[str, int] = {}
-        self.faiss_index: faiss.Index | None = None
-        self._index_dirty: bool = True
-        meta_exists = self.path.exists() and (self.path / "meta.yaml").exists()
-        if meta_exists:
-            self.load()
-        else:
-            os.makedirs(self.path, exist_ok=True)
-            if embedding_dim <= 0:
-                raise ValueError("embedding_dim must be > 0 for new stores")
-            self.meta = {
-                "version": 1,
-                "embedding_model": self.embedding_model,
-                "embedding_dim": self.embedding_dim,
-                "normalized": self.normalized,
-                "created_at": datetime.now(timezone.utc)
-                .isoformat()
-                .replace("+00:00", "Z"),
-                "updated_at": datetime.now(timezone.utc)
-                .isoformat()
-                .replace("+00:00", "Z"),
-            }
-            self.proto_vectors = np.zeros((0, self.embedding_dim), dtype=np.float32)
-            self.save()
-
-    # ------------------------------------------------------------------
-    def _meta_path(self) -> Path:
-        return self.path / "meta.yaml"
-
-    def _proto_json_path(self) -> Path:
-        return self.path / "belief_prototypes.json"
-
-    def _proto_vec_path(self) -> Path:
-        return self.path / "prototype_vectors.npy"
-
-    def _mem_jsonl_path(self) -> Path:
-        return self.path / "raw_memories.jsonl"
-
-    # ------------------------------------------------------------------
-    def _build_faiss_index(self) -> None:
-        """Rebuild the FAISS index from ``self.proto_vectors``."""
-        if self.proto_vectors is None or len(self.prototypes) == 0:
-            self.faiss_index = None
-            self._index_dirty = False
-            return
-
-        index = faiss.IndexFlatIP(self.embedding_dim)
-        index.add(self.proto_vectors.astype(np.float32))
-        self.faiss_index = index
-        self._index_dirty = False
-
-    # ------------------------------------------------------------------
-    def load(self) -> None:
-        if not self._meta_path().exists():
-            raise FileNotFoundError("meta.yaml missing")
-        self.meta = yaml.safe_load(self._meta_path().read_text())
-        self.embedding_model = str(self.meta.get("embedding_model", "unknown"))
-        self.embedding_dim = int(self.meta.get("embedding_dim", 0))
-        if self.embedding_dim <= 0:
-            raise EmbeddingDimensionMismatchError("embedding_dim missing or invalid")
-        self.normalized = bool(self.meta.get("normalized", False))
-        if not self.normalized:
-            raise ValueError("embeddings must be normalized")
-
-        if self._proto_vec_path().exists():
-            arr = np.load(self._proto_vec_path())
-            if arr.ndim != 2 or arr.shape[1] != self.embedding_dim:
-                raise EmbeddingDimensionMismatchError(
-                    f"{arr.shape[1]} vs {self.embedding_dim}"
-                )
-            self.proto_vectors = arr.astype(np.float32)
-        else:
-            self.proto_vectors = np.zeros((0, self.embedding_dim), dtype=np.float32)
-
-        if self._proto_json_path().exists():
-            with open(self._proto_json_path(), "r") as f:
-                data = json.load(f)
-            self.prototypes = [BeliefPrototype(**p) for p in data]
-            self.index = {p.prototype_id: p.vector_row_index for p in self.prototypes}
-        if self._mem_jsonl_path().exists():
-            with open(self._mem_jsonl_path(), "r") as f:
-                for line in f:
-                    line = line.strip()
-                    if not line:
-                        continue
-                    self.memories.append(RawMemory.model_validate_json(line))
-        self._index_dirty = True
-        self.faiss_index = None
-
-    def save(self) -> None:
-        self.meta["updated_at"] = (
-            datetime.now(timezone.utc).isoformat().replace("+00:00", "Z")
-        )
-        with open(self._meta_path(), "w") as f:
-            yaml.safe_dump(self.meta, f)
-        with open(self._proto_json_path(), "w") as f:
-            json.dump([p.model_dump(mode="json") for p in self.prototypes], f)
-        if self.proto_vectors is not None:
-            np.save(self._proto_vec_path(), self.proto_vectors)
-        with open(self._mem_jsonl_path(), "w") as f:
-            for mem in self.memories:
-                f.write(mem.model_dump_json() + "\n")
-        self._index_dirty = True
-
-    # ------------------------------------------------------------------
-    def add_prototype(self, proto: BeliefPrototype, vec: np.ndarray) -> None:
-        idx = len(self.prototypes)
-        proto.vector_row_index = idx
-        self.prototypes.append(proto)
-        if self.proto_vectors is None or self.proto_vectors.size == 0:
-            self.proto_vectors = vec.reshape(1, -1)
-        else:
-            if vec.ndim == 1:
-                vec = vec.reshape(1, -1)
-            self.proto_vectors = np.vstack([self.proto_vectors, vec])
-        self.index[proto.prototype_id] = idx
-        self._index_dirty = True
-
-    def update_prototype(
-        self,
-        proto_id: str,
-        new_vec: np.ndarray,
-        memory_id: str,
-        *,
-        alpha: float | None = None,
-    ) -> float:
-        """Update ``proto_id`` towards ``new_vec`` using EMA and return the vector change magnitude.
-
-        This method incrementally evolves the prototype representation, enabling long-term
-        learning from accumulated evidence.
-        """
-
-        idx = self.index[proto_id]
-        assert self.proto_vectors is not None
-        current = self.proto_vectors[idx]
-        proto = self.prototypes[idx]
-        if alpha is None:
-            alpha = 1.0 / (proto.strength + 1.0) if proto.strength > 0 else 1.0
-        updated = (1 - alpha) * current + alpha * new_vec
-        change = float(np.linalg.norm(updated - current))
-        if self.normalized:
-            norm = np.linalg.norm(updated) or 1.0
-            updated = updated / norm
-        self.proto_vectors[idx] = updated.astype(np.float32)
-        proto.last_updated_ts = datetime.now(timezone.utc).replace(microsecond=0)
-        proto.constituent_memory_ids.append(memory_id)
-        proto.strength += 1.0
-        self._index_dirty = True
-        return change
-
-    def find_nearest(self, vec: np.ndarray, k: int) -> List[Tuple[str, float]]:
-        """Return ``k`` nearest prototypes by cosine similarity."""
-
-        if self.proto_vectors is None or len(self.prototypes) == 0:
-            return []
-
-        if (
-            self.faiss_index is None
-            or self._index_dirty
-            or self.faiss_index.ntotal != len(self.prototypes)
-        ):
-            self._build_faiss_index()
-
-        query = vec.astype(np.float32).reshape(1, -1)
-        dists, idxs = self.faiss_index.search(query, min(k, self.faiss_index.ntotal))
-        results: List[Tuple[str, float]] = []
-        for idx, dist in zip(idxs[0], dists[0]):
-            if idx < 0:
-                continue
-            results.append((self.prototypes[int(idx)].prototype_id, float(dist)))
-        return results
-
-    def add_memory(self, memory: RawMemory) -> None:
-        self.memories.append(memory)
-
-    # ------------------------------------------------------------------
-    def _repr_html_(self) -> str:
-        """Return HTML summary for notebooks."""
-        meta_items = """<tr><th>Embedding model</th><td>{model}</td></tr>
-<tr><th>Dimension</th><td>{dim}</td></tr>
-<tr><th>Normalized</th><td>{norm}</td></tr>
-<tr><th>Prototypes</th><td>{pcount}</td></tr>
-<tr><th>Memories</th><td>{mcount}</td></tr>
-<tr><th>Path</th><td>{path}</td></tr>""".format(
-            model=self.embedding_model,
-            dim=self.embedding_dim,
-            norm=self.normalized,
-            pcount=len(self.prototypes),
-            mcount=len(self.memories),
-            path=self.path,
-        )
-        return f"<table>{meta_items}</table>"
