+++ tests/test_agent.py
@@ -6,7 +6,7 @@ import pytest
 from compact_memory import agent as ag
 from compact_memory.agent import Agent
 from compact_memory.embedding_pipeline import MockEncoder, _load_model, embed_text
-from compact_memory.json_npy_store import JsonNpyVectorStore
+from compact_memory.vector_store import InMemoryVectorStore
 from compact_memory.chunker import SentenceWindowChunker, Chunker
 from compact_memory.active_memory_manager import ActiveMemoryManager
 from compact_memory.prompt_budget import PromptBudget
@@ -15,12 +15,14 @@ from compact_memory.prompt_budget import PromptBudget
 @pytest.fixture(autouse=True)
 def use_mock_encoder(monkeypatch):
     enc = MockEncoder()
-    monkeypatch.setattr("compact_memory.embedding_pipeline._load_model", lambda *a, **k: enc)
+    monkeypatch.setattr(
+        "compact_memory.embedding_pipeline._load_model", lambda *a, **k: enc
+    )
     yield
 
 
 def test_duplicate_handling(tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
     agent.add_memory("alpha")
     dup = agent.add_memory("alpha")
@@ -30,7 +32,7 @@ def test_duplicate_handling(tmp_path):
 
 
 def test_snap_and_spawn(tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
     agent.add_memory("alpha")
     res = agent.add_memory("golf")[0]
@@ -42,7 +44,7 @@ def test_snap_and_spawn(tmp_path):
 
 
 def test_initial_summary(tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
     agent.add_memory("alpha bravo charlie")
     assert store.prototypes[0].summary_text != ""
@@ -50,7 +52,7 @@ def test_initial_summary(tmp_path):
 
 
 def test_summary_update(tmp_path, monkeypatch):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker(), update_summaries=True)
     agent.add_memory("alpha bravo")
     first = store.prototypes[0].summary_text
@@ -67,21 +69,8 @@ def test_summary_update(tmp_path, monkeypatch):
     assert "delta" in updated
 
 
-def test_persistence_roundtrip(tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
-    agent = Agent(store, chunker=SentenceWindowChunker())
-    agent.add_memory("alpha")
-    vec = embed_text(["alpha"])[0]
-    before = store.find_nearest(vec, k=1)[0]
-    store.save()
-    store2 = JsonNpyVectorStore(path=str(tmp_path))
-    after = store2.find_nearest(vec, k=1)[0]
-    assert before[0] == after[0]
-    assert abs(before[1] - after[1]) < 1e-6
-
-
 def test_receive_channel_ingest(tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
     res = agent.receive_channel_message("user", "hello there")
     assert res["action"] == "ingest"
@@ -90,7 +79,7 @@ def test_receive_channel_ingest(tmp_path):
 
 
 def test_receive_channel_query(monkeypatch, tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
     agent.add_memory("alpha bravo")
 
@@ -115,14 +104,16 @@ def test_receive_channel_query(monkeypatch, tmp_path):
 
 
 def test_process_conversational_turn_updates_manager(monkeypatch, tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
 
     class Dummy:
         def __init__(self, *a, **k):
             pass
 
-        tokenizer = staticmethod(lambda text, return_tensors=None: {"input_ids": [text.split()]})
+        tokenizer = staticmethod(
+            lambda text, return_tensors=None: {"input_ids": [text.split()]}
+        )
         model = type("M", (), {"config": type("C", (), {"n_positions": 50})()})()
         max_new_tokens = 10
 
@@ -140,7 +131,7 @@ def test_process_conversational_turn_updates_manager(monkeypatch, tmp_path):
 
 
 def test_prompt_budget_truncates_prompt(monkeypatch, tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     budget = PromptBudget(query=2, recent_history=0, older_history=0, ltm_snippets=0)
     agent = Agent(store, chunker=SentenceWindowChunker(), prompt_budget=budget)
 
@@ -148,7 +139,9 @@ def test_prompt_budget_truncates_prompt(monkeypatch, tmp_path):
         def __init__(self, *a, **k):
             pass
 
-        tokenizer = staticmethod(lambda text, return_tensors=None: {"input_ids": text.split()})
+        tokenizer = staticmethod(
+            lambda text, return_tensors=None: {"input_ids": text.split()}
+        )
         model = type("M", (), {"config": type("C", (), {"n_positions": 50})()})()
         max_new_tokens = 10
 
@@ -168,7 +161,7 @@ def test_prompt_budget_truncates_prompt(monkeypatch, tmp_path):
 
 
 def test_get_statistics_ephemeral_store(tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     store.path = None
     agent = Agent(store, chunker=SentenceWindowChunker())
     stats = agent.get_statistics()
@@ -183,7 +176,7 @@ class DummyChunker(Chunker):
 
 
 def test_reconfigure_chunker(tmp_path):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
     agent.add_memory("alpha")
     agent.chunker = DummyChunker()
@@ -193,7 +186,7 @@ def test_reconfigure_chunker(tmp_path):
 
 
 def test_reconfigure_similarity_threshold(tmp_path, monkeypatch):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
     agent.add_memory("alpha")
 
@@ -205,4 +198,3 @@ def test_reconfigure_similarity_threshold(tmp_path, monkeypatch):
     res = agent.add_memory("bravo")[0]
     assert res["spawned"] is False
     assert store.meta["tau"] == 0.5
-
