+++ tests/test_conflict.py
@@ -3,7 +3,7 @@ import json
 import pytest
 
 from compact_memory import Agent
-from compact_memory.json_npy_store import JsonNpyVectorStore
+from compact_memory.vector_store import InMemoryVectorStore
 from compact_memory.embedding_pipeline import MockEncoder, _load_model
 from compact_memory.chunker import SentenceWindowChunker
 
@@ -11,12 +11,14 @@ from compact_memory.chunker import SentenceWindowChunker
 @pytest.fixture(autouse=True)
 def use_mock_encoder(monkeypatch):
     enc = MockEncoder()
-    monkeypatch.setattr("compact_memory.embedding_pipeline._load_model", lambda *a, **k: enc)
+    monkeypatch.setattr(
+        "compact_memory.embedding_pipeline._load_model", lambda *a, **k: enc
+    )
     yield
 
 
 def test_negation_conflict_logging(tmp_path, monkeypatch):
-    store = JsonNpyVectorStore(path=str(tmp_path), embedding_model="mock", embedding_dim=MockEncoder.dim)
+    store = InMemoryVectorStore(embedding_dim=MockEncoder.dim, path=str(tmp_path))
     agent = Agent(store, chunker=SentenceWindowChunker())
     agent.add_memory("Alice is happy.")
     pid = store.prototypes[0].prototype_id
@@ -24,6 +26,8 @@ def test_negation_conflict_logging(tmp_path, monkeypatch):
     agent.add_memory("Alice is not happy.")
     log_path = tmp_path / "conflicts.jsonl"
     assert log_path.exists()
-    rows = [json.loads(l) for l in log_path.read_text().splitlines() if l.strip()]
+    rows = [
+        json.loads(line) for line in log_path.read_text().splitlines() if line.strip()
+    ]
     assert len(rows) == 1
     assert rows[0]["prototype_id"] == pid
